Add basic footer with social media links
There are a ton of styling updates needed here but this is the base.

I'd like to find some really cute logos that match. It's been challenging to find a Github logo that matches the other social media links. Might have to buy something for this since I'm coming up short on free ones.